Refactor score manager
